sentiment_service/requirements.txt
flask
nltk
requests
python-dotenv
vadersentiment